requirements.txt
# Data
pandas>=2.2
numpy>=1.26
pyarrow>=15.0
datasets>=2.19          # HuggingFace streaming download

# Machine Learning
scikit-learn>=1.4
xgboost>=2.0
joblib>=1.3

# API
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
httpx>=0.27             # async HTTP client used by FastAPI test client

# Database (Project 1 PostgreSQL)
psycopg[binary]>=3.1
sqlalchemy>=2.0
python-dotenv>=1.0

# EDA / Visualisation
matplotlib>=3.8
seaborn>=0.13
jupyter>=1.0
ipykernel>=6.29
nbformat>=5.9

# Testing
pytest>=8.0
